backend/requirements.txt
Flask
flask-cors
mysql-connector-python
bcrypt
python-dotenv
PyJWT